it's 6am and im still coding send help

+++ src/main/java/xyz/sunrose/wacky_wonders/TargetUtils.java
@@ -42,7 +42,7 @@ public class TargetUtils {
 	}
 
 	// yeet function
-	public static void knockback(Entity attacker, LivingEntity target, int strength) {
+	public static void knockback(Entity attacker, LivingEntity target, double strength) {
 		target.takeKnockback(
 				strength,
 				MathHelper.sin(attacker.getYaw() * (float) (Math.PI / 180.0)),
